factor out functionality related to bravais systems to new (sub-)package Bravais.jl
