+++ Project.toml
@@ -11,7 +11,7 @@ JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Meshing = "e6723b4c-ebff-59f1-b4b7-d97aa5274f73"
 PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
@@ -23,7 +23,8 @@ JLD2 = "0.2.4, 0.4"
 Meshing = "0.5"
 PrettyTables = "0.12, 1.0, 1.1"
 Requires = "1.0, 1.1"
-StaticArrays = "1.0, 1.1, 1.2"
+StaticArrays = "1.2.8"
+Reexport = "1.0, 1.1, 1.2"
 julia = "1.4"
 
 [extras]
